Commit 4

+++ SW/src/Cash.java
@@ -2,7 +2,7 @@ import java.util.Map;
 
 public class Cash implements Payment_behaviour{
 //msh 3arfin its maiteen mother
-public void pay(Map<Integer,Integer>user_card,int price) {
+public void pay(Map<Long,Integer>user_card,int price) {
 	
 		
 	}
