+++ SW/src/Payment_behaviour.java
@@ -2,5 +2,5 @@ import java.util.Map;
 
 public interface Payment_behaviour {
 	
-abstract public void pay(Map<Integer,Integer>mp,int price);
+abstract public void pay(Map<Long,Integer>mp,int price);
 }
